refactor(style): no fs for `code,kbd,samp` in fix
Remove normalizing of font-size for `code,kbd,samp` from fix.css since default style
does not affect the line-height and generally is an OK style
(setting line-height for these elements is still kept). Also, these are
elements that either way will be styled for a site if they are actively used.